Update _brand.yml
Updated the _brand.yml to what I previously had (by no means "the best"), but selfishly think this looks nice as a default.

The fonts come from Google Fonts rather than anything system related- has always worked for me in the past but wanted to flag this. Also as we said previously, I don't think we need to think of the brand guidelines as bible for us, so I think Lato and Merriweather are much nicer fonts to read technical work than Helvetica.